+++ src/Matrix_Ops/MMULdag.c
@@ -48,7 +48,7 @@ multabdag( GLU_complex a[ NCNC ] ,
   a[2] = conj( b[1] ) * c[0] + conj( b[3] ) * c[2] ;	\
   a[3] = conj( b[1] ) * c[1] + conj( b[3] ) * c[3] ;
 #else
-  int i , j , m ;
+  size_t i , j , m ;
   register GLU_complex sum ;
   register GLU_real REB , IMB , REC , IMC ;
   for( i = 0 ; i < NC ; i++ ) {
